RAG_Backend/requirements_web.txt
flask==2.3.3
openai==0.28.1
transformers==4.35.0
torch==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
python-dotenv==1.0.0
numpy==1.24.3 